Add support for patterndb v6 type-hints
Syslog-ng 4.0 added type information to attributes.  This break the XML
files created by the module because no data-type is output (so strings
are assumed) but parters like @NUMBER@ do not return a string anymore.

Detect the Puppet type used and select the appropriate type when
generating the XML specification for PatternDB.