Fix bootstrapping of syslog-ng
When applying a catalog on a node without syslog-ng installed, the
`syslog_ng_version` fact is unset, and `versioncmp` fail because it
cannot compare undefined values.

The issue is not triggered in CI because we provide values for all facts
in order to test only the patterndb code.

Rely on the `get` function form stdlib to get the fact value so that we
can provide a default if the fact is not set.  This default value is set
to pre-4.0.0 in order for the version of the parrtedb to be compatible.