Make testing and installing more reliable
If a pattern database test is failing, it is not installed, but on
subsequent run of Puppet the non-working pattern database is considered
unchanged by puppet so not tested again and the configuration converge
to a state where the updated pattern database is not installed.

Rework the pattern database testing and installation so that a failling
pattern database fail on each subsequent run of Puppet, so that errors
can be spotted more easily.

Fixes #42