Make testing and installing more reliable
If a pattern database is failing, it is not installed, but on subsequent
run of Puppet the pattern database is not tested again and the
configuration converge to a state where the expected pattern database is
not installed.

Rework the pattern database testing and installation so that a failling
pattern database fail on each subsequent run of Puppet, so that errors
can be spotted more easily.

Fixes #42